NightTransition is fundamentally different from DayTransition
Alright, here's a little scene that plays through a sunset for night scenes and shit. Ezpz, just an animation and controller attached to the directional light, looks pretty decent

Anyways, I've also added some more sounds and shit, including an interaction sound to start with. We'll start throwing that together.